pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.0.5",
    "langchain-community>=0.3.26",
    "langchain[google-genai]>=0.3.26",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
]
